Move RepeatOp workaround to a separate file and add a test for Repeat workaround.
