+++ lib/Dialect/TTNN/Transforms/CMakeLists.txt
@@ -5,7 +5,8 @@ add_mlir_dialect_library(MLIRTTNNTransforms
         TTNNDecomposeLayouts.cpp
         TTNNToCpp.cpp
         Workarounds/Decomposition/ReduceOpsRewritePattern.cpp
-        Workarounds/TTNNWorkarounds.cpp
+        Workarounds/Decomposition/RepeatOpRewritePattern.cpp
+	Workarounds/TTNNWorkarounds.cpp
 
         ADDITIONAL_HEADER_DIRS
         ${PROJECT_SOURCE_DIR}/include/ttmlir
